Extract inherents to separate module
