+++ src/author/runtime/tests.rs
@@ -17,6 +17,7 @@
 
 #![cfg(test)]
 
+use crate::verify::inherents;
 use core::iter;
 
 #[test]
@@ -68,9 +69,9 @@ fn block_building_works() {
             super::BlockBuild::ApplyExtrinsic(ext) => builder = ext.finish(),
             super::BlockBuild::ApplyExtrinsicResult { .. } => unreachable!(),
             super::BlockBuild::InherentExtrinsics(ext) => {
-                builder = ext.inject_inherents(super::InherentData {
+                builder = ext.inject_inherents(inherents::InherentData {
                     timestamp: 1234,
-                    consensus: super::InherentDataConsensus::Aura { slot_number: 1234 },
+                    consensus: inherents::InherentDataConsensus::Aura { slot_number: 1234 },
                 });
             }
             super::BlockBuild::StorageGet(get) => {
