+++ src/verify.rs
@@ -72,3 +72,4 @@ pub mod aura;
 pub mod babe;
 pub mod header_body;
 pub mod header_only;
+pub mod inherents;
